Fix: Send directions as float64 instead of float32
The Twist message declaration expects a float64 value and we don't have any automatic conversions yet

+++ examples/python-ros2-dataflow/control_node.py
@@ -22,6 +22,6 @@ for i in range(500):
             "direction",
             pa.array(
                 [random.random() + 1, 0, 0, 0, 0, (random.random() - 0.5) * 5],
-                type=pa.float32(),
+                type=pa.float64(),
             ),
         )
